refactor: try to make the code more readable
