+++ packages/@repo/command-extractor/src/client.ts
@@ -1,0 +1,8 @@
+import {createClient} from '@sanity/client'
+
+import {SANITY_CONFIG} from './constants.js'
+
+export const client = createClient({
+  ...SANITY_CONFIG,
+  token: process.env.SANITY_API_TOKEN,
+})
